update(25 Nov 2021)

+++ lib/modals/football_contest.dart
@@ -34,8 +34,8 @@ class FootballContestData {
     String status;
     String matchType;
     String totalGoals;
-    int betCoins;
-    int winningCoins;
+    double betCoins;
+    double winningCoins;
     String whoWon;
     String whoLose;
     double userlat;
@@ -54,8 +54,8 @@ class FootballContestData {
         status: json["status"],
         matchType: json["matchType"],
         totalGoals: json["totalGoals"],
-        betCoins: json["betCoins"],
-        winningCoins: json["winningCoins"],
+        betCoins: json["betCoins"].toDouble(),
+        winningCoins: json["winningCoins"].toDouble(),
         whoWon: json["whoWon"],
         whoLose: json["whoLose"],
         userlat: json["userlat"].toDouble(),
